fix: In Qt 6.10.1, QString::arg() no longer casts to integral types
https://doc.qt.io/qt-6/qstring.html#arg-2

+++ 3rdparty/karchive/src/kar.cpp
@@ -87,7 +87,7 @@ bool KAr::openArchive(QIODevice::OpenMode mode)
         return true;
     }
     if (mode != QIODevice::ReadOnly && mode != QIODevice::ReadWrite) {
-        setErrorString(tr("Unsupported mode %1").arg(mode));
+        setErrorString(tr("Unsupported mode %1").arg((int)mode));
         return false;
     }
